requirements.txt
aiofiles
discord.py
python-dotenv
pydantic
pydantic-settings
